Fix for #31: Moved test files and refactored imports in the registrar
