+++ src/main/java/com/devonfw/ide/sonarqube/common/impl/DevonSonarRegistrar.java
@@ -7,40 +7,40 @@ import org.sonar.plugins.java.api.CheckRegistrar;
 import org.sonar.plugins.java.api.JavaCheck;
 import org.sonarsource.api.sonarlint.SonarLintSide;
 
-import com.devonfw.ide.sonarqube.common.impl.check.DevonArchitecture3rdPartyDatatypeMappingsCheck;
-import com.devonfw.ide.sonarqube.common.impl.check.DevonArchitecture3rdPartyHibernateCheck;
-import com.devonfw.ide.sonarqube.common.impl.check.DevonArchitecture3rdPartyJpaCheck;
-import com.devonfw.ide.sonarqube.common.impl.check.DevonArchitecture3rdPartyMysemaCheck;
-import com.devonfw.ide.sonarqube.common.impl.check.DevonArchitecture3rdPartyObjectsCheck;
-import com.devonfw.ide.sonarqube.common.impl.check.DevonArchitecture3rdPartyTransactionalCheck;
-import com.devonfw.ide.sonarqube.common.impl.check.DevonArchitectureComponentDeclarationCheck;
-import com.devonfw.ide.sonarqube.common.impl.check.DevonArchitectureComponentDependencyCheck;
-import com.devonfw.ide.sonarqube.common.impl.check.DevonArchitectureLayerAny2ClientCheck;
-import com.devonfw.ide.sonarqube.common.impl.check.DevonArchitectureLayerBatch2DataaccessCheck;
-import com.devonfw.ide.sonarqube.common.impl.check.DevonArchitectureLayerBatch2Logic4ComponentCheck;
-import com.devonfw.ide.sonarqube.common.impl.check.DevonArchitectureLayerBatch2ServiceCheck;
-import com.devonfw.ide.sonarqube.common.impl.check.DevonArchitectureLayerClient2BatchCheck;
-import com.devonfw.ide.sonarqube.common.impl.check.DevonArchitectureLayerClient2DataaccessCheck;
-import com.devonfw.ide.sonarqube.common.impl.check.DevonArchitectureLayerClient2LogicCheck;
-import com.devonfw.ide.sonarqube.common.impl.check.DevonArchitectureLayerCommon2AnyCheck;
-import com.devonfw.ide.sonarqube.common.impl.check.DevonArchitectureLayerDataaccess2Dataaccess4ComponentCheck;
-import com.devonfw.ide.sonarqube.common.impl.check.DevonArchitectureLayerDataaccess2LogicCheck;
-import com.devonfw.ide.sonarqube.common.impl.check.DevonArchitectureLayerDataaccess2ServiceCheck;
-import com.devonfw.ide.sonarqube.common.impl.check.DevonArchitectureLayerLogic2Dataaccess4ComponentCheck;
-import com.devonfw.ide.sonarqube.common.impl.check.DevonArchitectureLayerLogic2ServiceCheck;
-import com.devonfw.ide.sonarqube.common.impl.check.DevonArchitectureLayerService2BatchCheck;
-import com.devonfw.ide.sonarqube.common.impl.check.DevonArchitectureLayerService2DataaccessCheck;
-import com.devonfw.ide.sonarqube.common.impl.check.DevonArchitectureLayerService2Logic4ComponentCheck;
-import com.devonfw.ide.sonarqube.common.impl.check.DevonArchitectureLayerService2Service4ComponentCheck;
-import com.devonfw.ide.sonarqube.common.impl.check.DevonArchitecturePackageCheck;
-import com.devonfw.ide.sonarqube.common.impl.check.DevonArchitectureScopeApi2Base4ComponentPartCheck;
-import com.devonfw.ide.sonarqube.common.impl.check.DevonArchitectureScopeApi2BaseCheck;
-import com.devonfw.ide.sonarqube.common.impl.check.DevonArchitectureScopeApi2ImplCheck;
-import com.devonfw.ide.sonarqube.common.impl.check.DevonArchitectureScopeBase2Base4ComponentPartCheck;
-import com.devonfw.ide.sonarqube.common.impl.check.DevonArchitectureScopeBase2Impl4ComponentPartCheck;
-import com.devonfw.ide.sonarqube.common.impl.check.DevonArchitectureScopeBase2ImplCheck;
-import com.devonfw.ide.sonarqube.common.impl.check.DevonArchitectureScopeImpl2Base4ComponentPartCheck;
-import com.devonfw.ide.sonarqube.common.impl.check.DevonArchitectureScopeImpl2Impl4ComponentPartCheck;
+import com.devonfw.ide.sonarqube.common.impl.check.component.DevonArchitectureComponentDeclarationCheck;
+import com.devonfw.ide.sonarqube.common.impl.check.component.DevonArchitectureComponentDependencyCheck;
+import com.devonfw.ide.sonarqube.common.impl.check.component.DevonArchitectureLayerBatch2Logic4ComponentCheck;
+import com.devonfw.ide.sonarqube.common.impl.check.component.DevonArchitectureLayerDataaccess2Dataaccess4ComponentCheck;
+import com.devonfw.ide.sonarqube.common.impl.check.component.DevonArchitectureLayerLogic2Dataaccess4ComponentCheck;
+import com.devonfw.ide.sonarqube.common.impl.check.component.DevonArchitectureLayerService2Logic4ComponentCheck;
+import com.devonfw.ide.sonarqube.common.impl.check.component.DevonArchitectureLayerService2Service4ComponentCheck;
+import com.devonfw.ide.sonarqube.common.impl.check.layer.DevonArchitectureLayerAny2ClientCheck;
+import com.devonfw.ide.sonarqube.common.impl.check.layer.DevonArchitectureLayerBatch2DataaccessCheck;
+import com.devonfw.ide.sonarqube.common.impl.check.layer.DevonArchitectureLayerBatch2ServiceCheck;
+import com.devonfw.ide.sonarqube.common.impl.check.layer.DevonArchitectureLayerClient2BatchCheck;
+import com.devonfw.ide.sonarqube.common.impl.check.layer.DevonArchitectureLayerClient2DataaccessCheck;
+import com.devonfw.ide.sonarqube.common.impl.check.layer.DevonArchitectureLayerClient2LogicCheck;
+import com.devonfw.ide.sonarqube.common.impl.check.layer.DevonArchitectureLayerCommon2AnyCheck;
+import com.devonfw.ide.sonarqube.common.impl.check.layer.DevonArchitectureLayerDataaccess2LogicCheck;
+import com.devonfw.ide.sonarqube.common.impl.check.layer.DevonArchitectureLayerDataaccess2ServiceCheck;
+import com.devonfw.ide.sonarqube.common.impl.check.layer.DevonArchitectureLayerLogic2ServiceCheck;
+import com.devonfw.ide.sonarqube.common.impl.check.layer.DevonArchitectureLayerService2BatchCheck;
+import com.devonfw.ide.sonarqube.common.impl.check.layer.DevonArchitectureLayerService2DataaccessCheck;
+import com.devonfw.ide.sonarqube.common.impl.check.packaging.DevonArchitecturePackageCheck;
+import com.devonfw.ide.sonarqube.common.impl.check.scope.DevonArchitectureScopeApi2Base4ComponentPartCheck;
+import com.devonfw.ide.sonarqube.common.impl.check.scope.DevonArchitectureScopeApi2BaseCheck;
+import com.devonfw.ide.sonarqube.common.impl.check.scope.DevonArchitectureScopeApi2ImplCheck;
+import com.devonfw.ide.sonarqube.common.impl.check.scope.DevonArchitectureScopeBase2Base4ComponentPartCheck;
+import com.devonfw.ide.sonarqube.common.impl.check.scope.DevonArchitectureScopeBase2Impl4ComponentPartCheck;
+import com.devonfw.ide.sonarqube.common.impl.check.scope.DevonArchitectureScopeBase2ImplCheck;
+import com.devonfw.ide.sonarqube.common.impl.check.scope.DevonArchitectureScopeImpl2Base4ComponentPartCheck;
+import com.devonfw.ide.sonarqube.common.impl.check.scope.DevonArchitectureScopeImpl2Impl4ComponentPartCheck;
+import com.devonfw.ide.sonarqube.common.impl.check.thirdparty.DevonArchitecture3rdPartyDatatypeMappingsCheck;
+import com.devonfw.ide.sonarqube.common.impl.check.thirdparty.DevonArchitecture3rdPartyHibernateCheck;
+import com.devonfw.ide.sonarqube.common.impl.check.thirdparty.DevonArchitecture3rdPartyJpaCheck;
+import com.devonfw.ide.sonarqube.common.impl.check.thirdparty.DevonArchitecture3rdPartyMysemaCheck;
+import com.devonfw.ide.sonarqube.common.impl.check.thirdparty.DevonArchitecture3rdPartyObjectsCheck;
+import com.devonfw.ide.sonarqube.common.impl.check.thirdparty.DevonArchitecture3rdPartyTransactionalCheck;
 
 /**
  * {@link CheckRegistrar} for this plugin.
