+++ src/test/files/component/DevonArchitectureComponentDeclarationCheck.java
