+++ src/test/files/component/DevonArchitectureLayerBatch2Logic4ComponentCheck.java
