+++ src/test/files/packaging/DevonArchitecturePackageCheck_NoScope.java
