+++ src/test/files/scope/DevonArchitectureScopeApi2Base4ComponentPartCheck.java
