+++ src/test/files/scope/DevonArchitectureScopeApi2BaseCheck.java
