+++ src/test/files/scope/DevonArchitectureScopeBase2Base4ComponentPartCheck.java
