+++ src/test/files/scope/DevonArchitectureScopeBase2Impl4ComponentPartCheck.java
