+++ src/test/files/scope/DevonArchitectureScopeBase2ImplCheck.java
