+++ src/test/files/scope/DevonArchitectureScopeImpl2Base4ComponentPartCheck.java
