+++ src/test/files/scope/DevonArchitectureScopeImpl2Impl4ComponentPartCheck.java
